+++ svrcore/policies/throttling.go
@@ -8,10 +8,10 @@ import (
 	"github.com/JeffreyRichter/svrcore"
 )
 
-func NewThrottlingPolicy(maxRequestsPerSecond int64) svrcore.Policy {
+func NewThrottlingPolicy(maxRequestsPerSecond int) svrcore.Policy {
 	requestPerSecond := newRateCounter(time.Second)
 	return func(ctx context.Context, r *svrcore.ReqRes) bool {
-		if requestPerSecond.Rate() >= int64(maxRequestsPerSecond) {
+		if requestPerSecond.Rate() >= maxRequestsPerSecond {
 			return r.WriteError(http.StatusTooManyRequests, nil, nil, "TooManyRequests", "Too many requests")
 		}
 		return r.Next(ctx)
